+++ src/app/_services/validation-service/validation.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ValidationService } from './validation.service';
+
+describe('ValidationService', () => {
+    let service: ValidationService;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({});
+        service = TestBed.inject(ValidationService);
+    });
+
+    it('should be created', () => {
+        expect(service).toBeTruthy();
+    });
+});
