--- src/app/services/validation-service/validations-service.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ValidationsServiceService } from './validations-service.service';
-
-describe('ValidationsServiceService', () => {
-  let service: ValidationsServiceService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ValidationsServiceService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
